+++ src/onegov.core/onegov/core/orm/types/__init__.py
@@ -2,5 +2,6 @@ from onegov.core.orm.types.hstore_type import HSTORE
 from onegov.core.orm.types.json_type import JSON
 from onegov.core.orm.types.utcdatetime_type import UTCDateTime
 from onegov.core.orm.types.uuid_type import UUID
+from onegov.core.orm.types.lowercase_text_type import LowercaseText
 
-__all__ = ['HSTORE', 'JSON', 'UTCDateTime', 'UUID']
+__all__ = ['HSTORE', 'JSON', 'UTCDateTime', 'UUID', 'LowercaseText']