+++ src/onegov.core/onegov/core/orm/types/lowercase_text_type.py
@@ -1,0 +1,15 @@
+from sqlalchemy.types import TypeDecorator, TEXT
+from sqlalchemy_utils.operators import CaseInsensitiveComparator
+
+
+class LowercaseText(TypeDecorator):
+    """ Text column that forces all text to be lowercase. """
+
+    impl = TEXT
+    omparator_factory = CaseInsensitiveComparator
+
+    def process_bind_param(self, value, dialect):
+        if value is not None:
+            return value.lower()
+
+        return value